[CARBONDATA-2304][Compaction] Prefetch rowbatch during compaction
Add a configuration to enable prefetch during compaction.

During compaction, carbondata will query on the segments and retrieve a row， then it will sort the rows and produce the final carbondata file.

Currently we find the poor performance in retrieving the rows, so adding prefetch for the rows will surely improve the compaction performance.

This closes #2133