[CARBONDATA-2428] Support flat folder for managed carbon table
Currently carbondata writing happens in fixed path tablepath/Fact/Part0/Segment_NUM folder and it is not same as hive/parquet folder structure. This PR makes all files written will be inside tablepath, it does not maintain any segment folder structure. Only for partition it adds the folder.

This feature can be controlled through a table property flat_folder. The default value of it is false.

This closes #2207