[CARBONDATA-2376] Improve Lucene datamap performance by eliminating blockid while writing and reading index
Problem:
Currently DataMap interface implementations use blockid and blockletid while writing index files, Actually blockid is not needed to store in index files as it only requires blockletid. So it adds more memory and disk size to write index files.

Solution:
Use taskname as index name to identify the indexname. And filter the blocklets directly by avoiding blockids.And pass the taskName as indexname to identify the blockid from blocletdatamap.

Corrected the implementations of LuceneDatamap, BloomFilterDataMap, CGDataMap, FGDataMap and MinMaxDataMap

This closes #2206