[CARBONDATA-2386]Fix Query performance issue with PreAggregate table
Problem: Query on pre aggregate table is consuming too much time.
Root cause: Time consumption to calculate size of selecting the smallest Pre-Aggregate table is approximately 76 seconds. This is index file is being read when segment file is present to compute the size of Pre-Aggregate table
Solution: Read table status and get the size of data file and index file for valid segments. For older segments were datasize and indexsize is not present calculate the size of store folder

This closes #2213