[CARBONDATA-2391] [Compaction Thread Leak] Thread leak in compaction operation if prefetch is enabled and compaction process is killed
Problem
Thread leak in compaction operation if prefetch is enabled and compaction process is killed

Analysis
During compaction if prefetch is enabled RawResultIterator launches an executor service for prefetching the data.
If compaction fails or the process is killed it can lead to thread leak due to executor service still in running state.

Fix
Close the executor service in task completion listener of compaction

This closes #2217