[CARBONDATA-2394] Setting segments in thread local space but was not getting reflected in the driver is fixed
Problem : from multiple thread if we are setting any property to thread level. It is affecting other also.
Analysis : information in carbonSessionInfo object is being corrupted inside CarbonEnv. so if new thread may be impacted by prev one.
Solution : create a new object and in set the user specified values in thread/session params to new object.

This closes #2223