[CARBONDATA-2417] SDK writer goes to infinite wait when consumer thread is dead
problem: SDK writer goes to infinite wait when consumer thread is dead

root cause: due to bad record when an exception happens at consumer thread
during write, this message is not reached producer (SDK writer).
So, SDK keeps writing data assuming consumer will consume it. But as a consumer is dead, queue becomes full and queue.put() will be blocked
forever.

Solution: When the consumer is dead, call writer.close() forcefully and
clear queue. so that blocking write will go an and skip writing next
batches. when writer.close() is called by user. Throw an exception that
write is failed

This closes #2251