[CARBONDATA-2426] Added fix for data mismatch after compaction on Partition with Pre-Aggregate tables
Problem: Partition directory is getting deleted when one on the segment is marked for delete and
another segment is loaded in the same partition.
In case of aggregate table we only have partition specs for the specified segments therefore
when the deleted segment is scanned against locationMap for stale partitions one of the
valid partition is considered as stale as we dont get partitionSpecs for all the segments.

Solution: Delete the index file instead of the partition directory and then delete the directory if it is empty.

This closes #2259