[CARBONDATA-2433][LUCENE]close the lucene index reader after every task and clean the resource and other functional issues
problem:

Lucene IndexReader opened during query is never closed. Which will impact performance and will lead to memory issues
Lucene index will not index the stop words like is, the etc. which may lead to wrong data with filter queries with text match
Solution:

Close the index reader once the task is completed.
Make it configurable to index the stop words in lucene indexing

This closes #2269