[CARBONDATA-2436] Block pruning problem post the carbon schema restructure.
Currently DataMap is pruning with segmentproperties from the 0th block of BlockletDataMap is not correct. As post restructure if the table is updated then all the block
will not have symmetric schema within the same segments.

Solution : It must be ensured the block could be pruned with the same schema.

This closes #2263