Problem: SDK Writer going into infinte loop in case of multi-threaded scenario
Analysis:
In multi-threaded scenarios where multiple instances of writer thread are trying to add a row to the RowBatch, addition to given
batch size cannot be ensured as addition process is not synchronized and it can lead to ArrayIndexOutOfBound Exception or data loss/mismatch issues
If multiple writer threads are adding the data to RowBatch and immediately after launching all the threads closeWriter is called, in that case we
don't know when all the data is finished writing by all the threads but we are returning immediately from close writer after setting the close flag
to true. This does not ensure complete processing of data.

Solution:
This API is meant for single threaded scenarios and is not thread safe. The same is updated in the interface docs.