[CARBONDATA-2467] sdk writer log shouldnot print null
issue :

Logger prints the threadname along with log. In case of sdk writer we are assigning null for dbname and table name and table name is used to assign the thread name in CarbonTableOutputFormat and DataWriterProcessStepImpl. which leads to add null in many logs.
Comparision of two array type was being wrong.
Solution :

assign temp value with current time in millisec to tablename instead of null.
ArrayType class donot contains equls method. Hence it goes to Object.equals() which will perform == operation to compare both the object which is wrong . so add equals method to ArrayType class .

This closes #2292