Problem: Problem inferring the complex type schema with boolean array type from the store created using SDK writer
Analysis:
When we create an external table and infer the schema from store created using SDK writer, the operation fails because of complex type field with
boolean array dataType. This is because during schema creation by SDK writer, for array type children a child with column name val is added.
While parsing the logic to append the parent name with child column name is missing for boolean type which is causing this problem.

Solution:
Handle the parsing for boolean type