[CARBONDATA-2482] Pass uuid while writing segment file if possible
Pass the uuid (segmentFileName) to the writeSegmentFile method file if possible.

Problem:
When the supporting tables depends on the segmentFileName of the main table, query is failing, as it is expected to be the same name as that of the main table. Mostly in case of merge index, segmentFile will be rewritten for that segment.

Solution:
Whenever the supporting table segmentFileName should be same as that main table, in that case we should pass the name as the UUID in the merge index flow instead of taking a new timestamp.

This closes #2307