[CARBONDATA-2578] fixed memory leak inside CarbonReader and handled failure case for creation of  multi reader for non-transactional table
Issue :
CarbonIterator inside CarbonRecordReader was keeping reference of RowBatch and it is not being
closed inside CarbonRecordReader. sort_column with measure was considering all the dimension
column along with given column. if creation of one CarbonReader for non transactional table
is failed then we are not able to create another CarbonReader.

Solution :
close() called inside hasNext() to clear the previous iterator before iterating over the next CarbonReader..
if sortcolumn is not empty and sortcolumnsList contains the fields check is added to finally in sortcolumn props.
Clear the datamap in catch block, if creation of CarbonReader is failed.

This closes #2362