[CARBONDATA-2528][MV] Fixed order by in mv and aggregation functions inside projection expressions are fixed
Problem:
Order by queries and the queries with functions like sum(a)+sum(b) are not working in MV.
Please check jira for more details.
Solution:
The queries which have projection functions like sum(a)+sum(b) cannot be incrementally loaded, so introduced a new internal DM property to avoid group by on the final query.

This closes #2453