[CARBONDATA-2608] SDK Support JSON data loading directly (without AVRO conversion)
What changes were proposed in this pull request?
Currently SDK Support JSON data loading only with AVRO.
So, converting json to avro record and avro to carbon object is a two
step process. Hence there is a need for a new carbonWriter that works
with Json without AVRO.
This PR implents that.

Highlights:
 # Works with just the json data and carbon schema.
 # Implements hadoop's FileInputFormat to create JsonInputFormat.
 # supports reading multiple json files in a folder.
 # supports reading json data in multiline with record identifier.
 # supports single row json read and write.
 # Handles bad records when loading json data

How was this patch tested?

Manual testing, and UTs are added in this PR.