[CARBONDATA-2617] Invalid tuple-id and block id getting formed for Non partition table
Problem
Invalid tuple and block id getting formed for non partition table

Analysis
While creating a partition table a segment file was written in the Metadata folder under table structure. This was introduced during
development of partition table feature. At that time segment file was written only for partition table and it was used to distinguish between
parition and non partition table in the code. But later the code was modified to write the segment file for both parititon and non partition
table and the code to distinguish partition and non partition table was not modified which is causing this incorrect formation of block and tuple id.

Fix
Modify the logic to distinguish partitioned and non partitioned table and the same has been handled in this PR.

This closes #2385