[CARBONDATA-2585][CARBONDATA-2586][Local Dictionary]Local dictionary support for alter table, preaggregate, varchar datatype, alter set and unset
What changes were proposed in this pull request?
In this PR,

local dictionary support is added for alter table, preaggregate, varChar datatype, alter table set and unset command
UTs are added for local dictionary load support
All the validations related to above features are taken care in this PR
How was this patch tested?

All the tests were executed in 3 node cluster.
UTs and SDV test cases are added in the same PR

This closes #2401