[CARBONDATA-2650][Datamap] Fix bugs in negative number of skipped blocklets
Currently in carbondata, default blocklet datamap will be used to prune
blocklets. Then other indexdatamap will be used.
But the other index datamap works for segment scope, which in some
scenarios, the size of pruned result will be bigger than that of default
datamap, thus causing negative number of skipped blocklets in explain
query output.

Here we add intersection after pruning. If the pruned result size is
zero, we will finish the pruning.

This closes #2410