Fix bugs in bloomfilter for dictionary/sort/date index columns
For dictionary column, carbon convert literal value to dict value, then
convert dict value to mdk value, at last it stores the mdk value as
internal value in carbonfile.
For sort column and date column, the value has also been encoded.

Here in bloomfilter datamap, we will index on the encoded data, that is
to say:
For dictionary/date column, we use the surrogate key as bloom index key;
For sort column and ordinary dimensions, we use the plain bytes as bloom
index key;
For measures, we convert the value to bytes and use it as the bloom
index key.

Changes are made:

1. FieldConverters were refactored to extract common value convert methods.
2. BloomQueryModel was optimized to support converting literal value to
internal value.
2. fix bugs for int/float/date/timestamp as bloom index column
3. fix bugs in dictionary/sort column as bloom index column
4. add tests
5. block (deferred) rebuild for bloom datamap (contains bugs that does
not fix in this commit, another PR has been raised)