[CARBONDATA-2676]Support local Dictionary for SDK Writer
Currently local dictionary is supported for managed table which is created using sql .
We it should be supported for SDK Writer also.
This PR contains
a. Interface to supply dictionary threshold & Boolean to enable dictionary.
b. DataLoading flow
c. Query Flow of Local dictionary (it will be done when normal table start supports for same)
d. External Table loading and Query .

This closes #2433