[CARBONDATA-2819] Fixed cannot drop preagg datamap on table which has other type datamaps
As we know, carbon now write preagg datamap info to main table schema and write other type datamap info like bloom or lucene to .dmschema file in _system folder. When dropping datamap, we should both check the .dmschema files and main table info if has the datamap which to be deleted.

This closes #2641