[CARBONDATA-2729][file-format] Schema Compatibility problem between version 1.3.0 and 1.4.0
Problem:
In TableSchema the field Name schemaEvaluation is changed to schemaEvoluation and in DataMapSchema field name className is changed to providerName.
Due to this current Carbon version & Version 1.4.0 is not able to properly de-serialize the schema created with older Carbon Version.
Solution
To de-serialize old schema having different field name the multiple values for SerializedName annotation can be used.
For refrence please refer the gson PR#699.

This closes #2498