[CARBONDATA-2606][Complex DataType Enhancements] Fixed Projection Pushdown when Select filter contains Struct column.
Problem:
If Select filter contains Struct Column which is not in Projection list,
then only null value is stored for struct column given in filter and select query result is null.
Solution:
Pushdown Parent column of corresponding struct type if any struct column is present in Filter list.

This closes #2439