[CARBONDATA-2714] Support merge index files for the segment
Problem :
The first-time query of carbon becomes very slow. It is because of reading many small carbonindex files and cache to the driver at the first time.
Many carbonindex files are created in below case
Loading data in large cluster
For example, if the cluster size is 100 nodes then for each load 100 index files are created per segment. So after 100 loads, the number of carbonindex files becomes 10000. .
It will be slower to read all the files from the driver since a lot of namenode calls and IO operations.
Solution :
Merge the carbonindex files in two levels.so that we can reduce the IO calls to namenode and improves the read performance.
Merge within a segment.
Merge the carbonindex files to single file immediately after load completes within the segment. It would be named as a .carbonindexmerge file. It is actually not a true data merging but a simple file merge. So that the current structure of carbonindex files does not change. While reading we just read one file instead of many carbonindex files within the segment.

This closes #2482