[CARBONDATA-2606][Complex DataType Enhancements]Fix Null result if First two  Projection column have same parent and third column has different Parent Struct
Problem:
When multiple columns are there,then the first child elements is only going
to make parent Object Array. For all other cases it should be null.
For e.g. a : <b,c,d>. here as 'a' is the parent column and b, c, d are
child columns during traversal when we encounter the first element in
list i.e. column 'b','a' will be completely filled. In case when column
'c' and 'd' encountered then only place null in the output.
Hence, as Null is placed in the output, Select result is Null if First
two Projection column have same parent and third column has different parent Struct column.
Solution: Place null in the end of output

This closes #2489