[CARBONDATA-2648] Fixed NPE issue with legacy store when CACHE_LEVEL is Blocklet
Things done as part of this PR:

Fixed Null pointer exception when store is of <= 1.1 version and DataMap is of type BlockletDataMap.
Added clearing of SegmentProperties cache holder from executor
Problem 1:
Null pointer exception thrown when store is of <= 1.1 version and DataMap is of type BlockletDataMap.

Analysis:
In BlcokletDataMap schema is created to consider blockletInfo while adding to unsafe but in case of legacy store we dont weite the blocklet Info. This lead to Null pointer exception while calculating the size of row during adding to unsafe.

Solution
For legacy store always call super class methods which is BlockDataMap.

This closes #2499