[CABONDATA-2741]Fix for filling measure column data in wrong order by not skipping non existing measure column in case of restructure
Problem: In case of restructure measure ordinal list in class RestructureUtil only maintains the existing orders but filling data in method fillResultToColumnarBatch of class DictionaryBasedVectorResultCollector doesnot ignores the non existing columns in measureColumnInfo.
so, before it was trying to read the data of existing column in non exsting columns maked datatype mismatch.
Solution: Making syncup for both measureColumnInfo with measure ordinals.

This closes #2507