[CARBONDATA-2746][BloomDataMap] Fix bug for getting datamap file when table has multiple datamaps
Currently, if table has multiple bloom datamap and carbon is set to use distributed datamap, query will throw an exception when accessing the index file, because carbon gets all the datamaps but sets them with same datamap schema. The error is appeared when getting the full path of bloom index by concating index directory and index column. This PR fix this problem by filter the index directories of target datamap when using distributed datamap.

Test shows that lucene is not affected by this. On the other hand, lucene gets wrong result if we apply this filter

This closes #2512