[CARBONDATA-2751] Fixed Thread leak issue in data loading and Compatibility issue
### Problem:

Thread leak when user is killing data loading process from UI

NPE when user is querying old store.

### Solution

When carbondata file writing is in progress during data loading and user is killing it from UI Producer and consumer thread are not getting shutdown. Need to handle the same in close method

Old store (V1/V2) does not have datachunk3 object so while filling the local dictionary it is checking whether local dictionary is present in datachunk3 or not but datachunk3 null check is missing

This closes #2521