[HOTFIX] Improved BlockDataMap caching performance during first time query
Things done as part of this PR

Created taskSumamry and FileFooterEntry schema once and stored in member variable.
Everytime creation of schema was a costly operation as time to prune dataMaps increased because of that.

Used TreeMap instead of HashMap while adding the complete file path and data to the map diring merge file read.
Using TreeMap improved the map filling performance by 10 sec for 1200 entries.

This closes #2531