[CARBONDATA-2753] Fix Compatibility issues on index Files with 1.3 store
Problem:
Currently,in the segmentFile we are writing the index files list in files field, only if it exists, otherwise it will be empty(in case if it is
merged to merge index file). But in the old store, we were writing both the files and mergeFileName fields even if the index files are merged.

Solution:
While querying we have to check the physical existence of the index files listed in the files field. If it physically exists, then we have to consider that.

This closes #2534