[CARBONDATA-2649] Handled executor min/max pruning when filter column in not cached in driver for CACHE_LEVEL=BLOCKLET
Things handled as part of this PR:

Modified code to use min/max in executor pruning for Blocklet dataMap when filter column min/max is not cached in driver. When column to be cached in driver are specified and CACHE_LEVEL = BLOCKLET, then executor min/max pruning was not happening which can increase the query time.

Removed unwanted addition of schemaEvolutionEntry to schema on Alter SET and UNSET table properties

This closes #2540