[CARBONDATA-2778]Fixed bug when select after delete and cleanup is showing empty records
Problem: In case if delete operation when it is found that the data being deleted is leading to a state where one complete block data is getting
deleted. In that case the status if that block is marked for delete and during the next delete operation run the block is deleted along with its
carbonIndex file. The problem arises due to deletion of carbonIndex file because for multiple blocks there can be one carbonIndex file as one
carbonIndex file represents one task.

Solution: Do not delete the carbondata and carbonIndex file. After compaction it will automatically take care of deleting the stale data and stale segments.

This closes #2548