optimize table pruning info for pruning by segment
In previous implementation, table pruning is performed once for all
segments. In this case, the pruning info is updated after default/CG/FG
datamap pruning.

Now we want the table pruning to be performed segment by segment, so in
this case, the pruning info will be updated after default/CG/FG datamap
pruning for each segment.

This means that we need to accumulate the pruning info during pruning.