[CARBONDATA-2807] Fixed data load performance issue in Intermediate merger When number of records are high
Problem: Data Loading is taking more time when number of records are high.
Root cause: As number of records are high intermediate merger is taking more time.
Solution: Checking the number of files present in file list is done is synchronized block because of this
each intermediate request is taking sometime and when number of records are high it impacting overall data loading performance

This closes #2588