[CARBONDATA-2813] Fixed code to get data size from LoadDetails if size is written there
Problem:
In 1.3.x when index files are merged to form mergeindex file a mapping of which index files if merged to which mergeindex is kept in the segments
file. In 1.4.x both the index and merge index files are scanned to calculate the size of segments for major compaction. As the index file was
deleted in the 1.3.x store therefore in 1.4.x it was throwing "Unable to get File status exception".

Solution:
Try to the size of the segments from LoadMetadataDetails. If not present then try to read the size from index files.

This closes #2600