[CARBONDATA-2817]Thread Leak in Update and in No sort flow
Issue :- After Update Command is finished , Loading threads are not getting stopped.

Root Cause :-

In Update flow DataLoadExecutor 's close method is not called so all Executors services are not closed.
In Exceptions are not handled property in AFDW class's closeExecutorService() which is cuasing Thread leak if Job is killed from SparkUI..
Solution :-

Add Task Completion Listener and call close method of DataLoadExecutor to it .
Handle Exception in closeExecutor Service so that all Writer steps Threads can be closed.

This closes #2606