[HOTFIX]Fixed int overflow and comparison gone wrong during blocklet min/max
Problem: During calculating min/max for blocklet, it needs to calculate from all the pages. During that comparison,
it is typecasting to int and overflows, so there is a chance the negative becomes positive and positive become negative.
That's why min max of long comes wrong for bigger values.
Solution: Don't typecast directly, instead check first the negative or positive and then return.