[CARBONDATA-2845][BloomDataMap] Merge bloom index files of multi-shards for each index column
Currently a bloom index file will be generated per task per load, the query performance will be bad if we have many segments. Main datamap already faced this problem before.
Here we want to merge the bloom index file in segment scope just like the main datamap do.

In this PR, we add an event listener to trigger action of merging bloom index files when datamap is built/rebuild.
File merging process uses an added RDD to work and write merged files by Thrift.
It is compatible to query on index files before/after merged. File mergeShard.inprogress is created as flag when the merge files is not ready so that query won't be disturbed when merging

This closes #2624