[HOTFIX] Support TableProperties Map API for SDK
Currently SDK supports load options as map input. But table properties is not a map.
So this PR supports a API that can take already supported table properties as map.

This is will help for easy configuration for end user of SDK.
Also later if new table properties added.
No need to create separate API to support that.

This closes #2651