[CARBONDATA-2925]Wrong data displayed for spark file format if carbon has more blocklets
Issue:- if Carbon file has multiple blocklet ,in select query wrong data displayed.
Root Cause :- it is showing records of only for 1st Blocklet and other blocklet in that block is getting skipped. This is because default blocklet is 0 and CarbonFileformat create blockletInfo with default configuration (not changed blockletID).
Solution :- Set default blockletID to -1 so that all blocklets are considered.
refer org.apache.carbondata.core.scan.executor.impl.AbstractQueryExecutor#readAndFillBlockletInfo

This closes #2703